--- src/Chapter01.Tests/Listing01.02C.Tests.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_02C.Tests
-{
-    [TestClass]
-    public class HelloWorldTests
-    {
-        [TestMethod]
-        public void Main_InigoHello()
-        {
-            const string expected =
-                @"Down
-Side
-Up";
-
-            IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, HelloWorld.Main);
-        }
-    }
-}
