--- src/Chapter01.Tests/Listing01.02E.Tests.cs
@@ -1,18 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_02E.Tests
-{
-    [TestClass]
-    public class HelloWorldTests
-    {
-        [TestMethod]
-        public void Main_InigoHello()
-        {
-            const string expected = 
-                @"Hello. My name is Inigo Montoya.";
-
-            IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, Program.Main);
-        }
-    }
-}
