+++ src/Chapter01.Tests/Listing01.04.Tests.cs
@@ -1,17 +1,20 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_04.Tests
 {
     [TestClass]
-    public class ProgramTests
+    public class HelloWorldTests
     {
         [TestMethod]
         public void Main_InigoHello()
         {
-            const string expected = @"Hello, My name is Inigo Montoya";
+            const string expected =
+                @"Down
+Side
+Up";
 
             IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, Program.Main);
+                expected, HelloWorld.Main);
         }
     }
 }
