+++ src/Chapter01.Tests/Listing01.08.Tests.cs
@@ -1,17 +1,17 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_08.Tests
 {
     [TestClass]
-    public class HelloWorldTests
+    public class ProgramTests
     {
         [TestMethod]
-        public void Main_HelloToInigo()
+        public void Main_InigoHello()
         {
-            const string expected = @"Hello Inigo Montoya";
+            const string expected = @"Hello, My name is Inigo Montoya";
 
             IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, HelloWorld.Main);
+                expected, Program.Main);
         }
     }
 }
