+++ src/Chapter01.Tests/Listing01.12.Tests.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_12.Tests
 {
@@ -6,15 +6,12 @@ namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_12.Tests
     public class MiracleMaxTests
     {
         [TestMethod]
-        public void Main_StormCastleLong()
+        public void Main_StormCastleShort()
         {
-            const string expected =
-@"Have fun storming the castle!
-Think it will work?
-It would take a miracle.";
+            const string expected = @"Have fun storming the castle!";
 
             IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, StormingTheCastle.Main);
+                expected, MiracleMax.Main);
         }
     }
 }
