+++ src/Chapter01.Tests/Listing01.13.Tests.cs
@@ -3,16 +3,16 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_13.Tests
 {
     [TestClass]
-    public class Listing01_13_Tests
+    public class HelloWorldTests
     {
         [TestMethod]
-        public void Main_ItWouldTakeAMiracle_OutputWrittenToConsole()
+        public void Main_InigoMantra_ProperOutput()
         {
             const string expected =
-@"It would take a miracle.";
+@"My name is Inigo Montoya.
+You killed my father....";
 
-            IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, StormingTheCastle.Main);
+            IntelliTect.TestTools.Console.ConsoleAssert.Expect(expected, HelloWorld.Main);
         }
     }
 }
