+++ src/Chapter01.Tests/Listing01.15.Tests.cs
@@ -1,17 +1,18 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_15.Tests
 {
     [TestClass]
-    public class ProgramTests
+    public class Listing01_13_Tests
     {
         [TestMethod]
-        public void Main_InputOne_WriteOne()
+        public void Main_ItWouldTakeAMiracle_OutputWrittenToConsole()
         {
-            const string expected = @"1<<1>>";
+            const string expected =
+@"It would take a miracle.";
 
             IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, Program.Main);
+                expected, StormingTheCastle.Main);
         }
     }
 }
