+++ src/Chapter01.Tests/Listing01.16.Tests.cs
@@ -1,18 +1,18 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_16.Tests
 {
     [TestClass]
-    public class ProgramTests
+    public class HeyYouTests
     {
         [TestMethod]
-        public void Main_InputInigoMontoya_WriteFirstLast()
+        public void Main_InputInigoMontoya_WriteNothing()
         {
             const string expected =
 @"Hey you!
 Enter your first name: <<Inigo
 >>Enter your last name: <<Montoya
->>Your full name is Inigo Montoya.";
+>>";
 
             IntelliTect.TestTools.Console.ConsoleAssert.Expect(
                 expected, HeyYou.Main);
