+++ src/Chapter01.Tests/Listing01.18.Tests.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_18.Tests
 {
@@ -6,16 +6,16 @@ namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_18.Tests
     public class ProgramTests
     {
         [TestMethod]
-        public void Main_InputInigoMontoya_WriteLastFirst()
+        public void Main_InputInigoMontoya_WriteFirstLast()
         {
             const string expected =
 @"Hey you!
 Enter your first name: <<Inigo
 >>Enter your last name: <<Montoya
->>Your full name is Montoya, Inigo.";
+>>Your full name is Inigo Montoya.";
 
             IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, Program.Main);
+                expected, HeyYou.Main);
         }
     }
 }
