+++ src/Chapter01.Tests/Listing01.21.Tests.cs
@@ -1,21 +1,21 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_19.Tests
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_21.Tests
 {
     [TestClass]
     public class ProgramTests
     {
         [TestMethod]
-        public void Main_InputInigoMontoya_WriteFirstLast()
+        public void Main_InputInigoMontoya_WriteLastFirst()
         {
             const string expected =
 @"Hey you!
 Enter your first name: <<Inigo
 >>Enter your last name: <<Montoya
->>Your full name is Inigo Montoya.";
+>>Your full name is Montoya, Inigo.";
 
             IntelliTect.TestTools.Console.ConsoleAssert.Expect(
-                expected, HeyYou.Main);
+                expected, CommentSamples.Main);
         }
     }
 }
