+++ src/Chapter01/Listing01.03.MultipleStatementsOneOneLine.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_02B
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_03
 {
     public class HelloWorld
     {
