+++ src/Chapter01/Listing01.06.HelloWorldInC#.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_02E
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_06
 {
     #region INCLUDE
     public class Program
