+++ src/Chapter01/Listing01.07.BasicClassDeclaration.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_03
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_07
 {
     #region INCLUDE
     public class HelloWorld
