+++ src/Chapter01/Listing01.09.TheMainMethodWithParametersAndAReturn.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_05
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_09
 {
     public class Program
     {
