+++ src/Chapter01/Listing01.10.NoIndentationFormatting.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_08
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_10
 {
 #region INCLUDE
 public class HelloWorld
