+++ src/Chapter01/Listing01.11.RemovingWhitespace.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_09;
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_11
 
 #region INCLUDE
 public class HelloWorld{public static void Main()
