+++ src/Chapter01/Listing01.13.DeclaringTwoVariablesWithinOneStatement.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_11
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_13
 {
     public class HelloWorld
     {
