+++ src/Chapter01/Listing01.16.UsingSystemConsoleReadLine.cs
@@ -15,11 +15,6 @@ namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_16
 
             Console.Write("Enter your last name: ");
             lastName = Console.ReadLine();
-
-            Console.WriteLine(
-            #region HIGHLIGHT
-                $"Your full name is { firstName } { lastName }.");
-            #endregion HIGHLIGHT
         }
     }
     #endregion INCLUDE
