+++ src/Chapter01/Listing01.17.UsingSystemConsoleRead.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_15
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_17
 {
     public class Program
     {
