+++ src/Chapter01/Listing01.19.FormattingUsingCompositeFormatting.cs
@@ -1,4 +1,4 @@
-namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_14
+namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_19
 {
     #region INCLUDE
     public class HeyYou
@@ -13,8 +13,13 @@ namespace AddisonWesley.Michaelis.EssentialCSharp.Chapter01.Listing01_14
             Console.Write("Enter your first name: ");
             firstName = Console.ReadLine();
 
+            #region HIGHLIGHT
             Console.Write("Enter your last name: ");
             lastName = Console.ReadLine();
+            #endregion HIGHLIGHT
+
+            Console.WriteLine(
+                "Your full name is {0} {1}.", firstName, lastName);
         }
     }
     #endregion INCLUDE
